Move androguard installing to tool codemagic-cli-tools
